backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-jsonapi==7.0.0
djangorestframework-simplejwt==5.3.1
inflection==0.5.1
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
